feat: Add protocol support for extensions (#716)
# Description

This updates the protocol to support Extensions. Extension profiles are
documents that describe modifications or additions to the core A2A
profile. AgentCards can declare supported extensions via
`capabilities.extensions`. This declaration can include additional
information via `capabilities.extensions.params`. Clients can use this
as a basis for extension negotiation: if a client supports an extension
that the server supports, the client can then follow the protocol
defined by the extension. This can include, but is not specifically
limited to:

1. Requiring specific structure of messages, such as following a
specific schema for a DataPart.
2. Providing additional information in request/response Message and
Artifact objects via their `metadata` field.
3. Exposing additional methods beyond the core A2A protocol methods.

Messages and Artifacts can now include an indication of which extensions
are present in their content. I do not yet know if extension indications
are necessary on other data structures (such as `MessageSendParams` or
`TaskStatus`). I'd say extension indications are not strictly necessary
at all: it's possible to infer that an extension was followed by
inspecting the message content itself (are the metadata fields you
expect present, does the DataPart follow the schema you expect, etc).
However, they are useful for validation and simplifying extension
implementation logic: "you said you followed this extension, so I'll
validate fields are present and then perform my processing".

Fixes #585

---------

Co-authored-by: Holt Skinner <13262395+holtskinner@users.noreply.github.com>