docs: fix links on documentation site (#960)
There are a few links in the spec doc that are clearly meant to link to
the github repo as the files they link to are outside the docs folder.
This fixes them so they work on the generated site hosted at
https://a2a-protocol.org/

Ideally we would use relative links so that when reading the markdown
file in your IDE the links take you to the corresponding file within the
repo. I tried setting up a redirect in mkdocs.yml but the locally
running website kept downloading the file from github instead of
navigating to it. There is probably some plugin that will make this work
however for now I would like to simply fix the docs on the website.