fix(docs)!:  update specification to clarify web hook certificate verification (#977)
# Webhook Certificate Checks

## User Benefit
While it is specified that Clients SHOULD verify Server identity through
the certificate, it is not currently explicit that Servers also have
this responsibility when delivering an update to a web hook URL. It
would be beneficial to users if uniform requirements on identity
verification were explicitly specified. Having a variation in
requirements that depends on whether responses are delivered
synchronously or asynchronously is surprising and seems unintentional.
The risk involved in not confirming the web hook certificate includes a
potential breach of confidentiality of response payload.

## This Change
This PR proposes an additional statement making explicit the A2A Server
responsibility when delivering asynchronous payload updates to
authenticate the destination certificate.

## Compatibility
In theory, a previously fully conforming implementation that did not
perform a check on a web hook server certificate would not meet the new
SHOULD statement. I couldn't locate information about what this implies
eg. with respect to versioning.
In practice, it seems at least somewhat unlikely that implementations
does this because verification of server certificates is the default
behavior across many different software toolsets for some time.

Closes #978